requirements.txt
dashscope>=1.25.0
requests>=2.30.0
Pillow>=10.0.0
